Created and modified index.html
This file is the root, from here we can travel through different recipes, doing this
we show the usage of links and travel through folders using hatml.

Created and modified recipe1.html

This is the first recipe, we added an image to learn how to use the labbel <img>, also show
the usage of links going to the root file.